make Router standalone component
