+++ src/components.js
@@ -10,6 +10,7 @@ export {default as NodeLink} from './NodeLink'
 export {default as Outline} from './Outline'
 export {default as PackageHeader} from './PackageHeader'
 export {default as RouteMatch} from './RouteMatch'
+export {default as Router} from './Router'
 export {default as Section} from './Section'
 export {default as SideNav} from './SideNav'
 export {default as StatusLabel} from './StatusLabel'
